requirements.txt
addict==2.4.0
ansiterm==1.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
brotlipy==0.7.0
certifi 
cffi 
charset-normalizer 
click==8.1.3
coloredlogs==15.0.1
comm==0.1.3
ConfigArgParse==1.5.3
contourpy==1.0.7
cryptography 
cycler==0.11.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock 
Flask==2.3.1
flatbuffers==23.3.3
fonttools==4.39.3
gmpy2 
graph-tools==1.11
humanfriendly==10.0
idna 
imageio==2.27.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2 
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe 
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
mpmath==1.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx 
numpy
onnx==1.13.1
onnxruntime==1.14.1
open3d==0.17.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
perlcompat==1.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.4.0
plotly==5.14.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.6
pycparser
Pygments==2.15.1
pyOpenSSL 
pyparsing==3.0.9
pypng==0.20220715.0
pyquaternion==0.9.9
pyrsistent==0.19.3
PySocks 
Pytess==1.0.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
requests 
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
segment-anything 
six 
stack-data==0.6.2
sympy 
tbdump==1.4
tenacity==8.2.2
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions 
tzdata==2023.3
urllib3 
wcwidth==0.2.6
Werkzeug==2.3.0
widgetsnbextension==4.0.7
zipp==3.15.0
